+++ tests/core/test_utils.py
@@ -7,15 +7,6 @@ from rasa.core import utils
 from rasa.core.utils import replace_floats_with_decimals
 
 
-@pytest.fixture(scope="session")
-def loop():
-    loop = asyncio.new_event_loop()
-    asyncio.set_event_loop(loop)
-    loop = rasa.utils.io.enable_async_loop_debugging(loop)
-    yield loop
-    loop.close()
-
-
 def test_is_int():
     assert utils.is_int(1)
     assert utils.is_int(1.0)