+++ CHANGELOG.rst
@@ -24,6 +24,7 @@ Added
 Changed
 -------
 - Unknown sections in markdown format (NLU data) are not ignored anymore, but instead an error is raised.
+- Tests can now be run in parallel
 
 Removed
 -------